#73 restructure constraints and constraint results + tests
- get rid of ConstraintTest and ConstraintResultTest
- split tests in one test per constraint
- put all case classes related to one constraint into one file
