+++ src/main/scala/de/frosner/ddq/constraints/ConditionalColumnConstraint.scala
@@ -16,3 +16,18 @@ case class ConditionalColumnConstraint(statement: Column, implication: Column) e
   }
 
 }
+
+case class ConditionalColumnConstraintResult(constraint: ConditionalColumnConstraint,
+                                             violatingRows: Long,
+                                             status: ConstraintStatus) extends ConstraintResult[ConditionalColumnConstraint] {
+
+  val message: String = {
+    val constraintString = s"${constraint.statement} -> ${constraint.implication}"
+    val pluralS = if (violatingRows == 1) "" else "s"
+    status match {
+      case ConstraintSuccess => s"Constraint $constraintString is satisfied."
+      case ConstraintFailure => s"$violatingRows row$pluralS did not satisfy constraint $constraintString."
+    }
+  }
+
+}
