+++ src/main/scala/de/frosner/ddq/constraints/NumberOfRowsConstraint.scala
@@ -27,3 +27,17 @@ object NumberOfRowsConstraint {
   }
 
 }
+
+case class NumberOfRowsConstraintResult(constraint: NumberOfRowsConstraint,
+                                        actual: Long,
+                                        status: ConstraintStatus) extends ConstraintResult[NumberOfRowsConstraint] {
+
+  val message: String = {
+    val expected = constraint.expected
+    status match {
+      case ConstraintSuccess => s"The number of rows satisfies $expected."
+      case ConstraintFailure => s"The actual number of rows $actual does not satisfy $expected."
+    }
+  }
+
+}
