+++ src/main/scala/de/frosner/ddq/constraints/StringColumnConstraint.scala
@@ -16,3 +16,19 @@ case class StringColumnConstraint(constraintString: String) extends Constraint {
   }
 
 }
+
+case class StringColumnConstraintResult(constraint: StringColumnConstraint,
+                                        violatingRows: Long,
+                                        status: ConstraintStatus) extends ConstraintResult[StringColumnConstraint] {
+
+  val message: String = {
+    val constraintString = constraint.constraintString
+    val pluralS = if (violatingRows == 1) "" else "s"
+    status match {
+      case ConstraintSuccess => s"Constraint $constraintString is satisfied."
+      case ConstraintFailure => s"$violatingRows row$pluralS did not satisfy constraint $constraintString."
+    }
+  }
+
+}
+
