+++ src/main/scala/de/frosner/ddq/constraints/UniqueKeyConstraint.scala
@@ -3,6 +3,7 @@ package de.frosner.ddq.constraints
 import org.apache.spark.sql.{DataFrame, Column}
 
 case class UniqueKeyConstraint(columnNames: Seq[String]) extends Constraint {
+
   require(columnNames.nonEmpty)
 
   val fun = (df: DataFrame) => {
@@ -10,4 +11,24 @@ case class UniqueKeyConstraint(columnNames: Seq[String]) extends Constraint {
     val nonUniqueRows = df.groupBy(columns: _*).count.filter(new Column("count") > 1).count
     UniqueKeyConstraintResult(this, nonUniqueRows, if (nonUniqueRows == 0) ConstraintSuccess else ConstraintFailure)
   }
+
+}
+
+case class UniqueKeyConstraintResult(constraint: UniqueKeyConstraint,
+                                     numNonUniqueTuples: Long,
+                                     status: ConstraintStatus) extends ConstraintResult[UniqueKeyConstraint] {
+
+  val message: String = {
+    val columnNames = constraint.columnNames
+    val columnsString = columnNames.mkString(", ")
+    val isPlural = columnNames.length > 1
+    val columnNoun = "Column" + (if (isPlural) "s" else "")
+    val columnVerb = if (isPlural) "are" else "is"
+    val pluralS = if (numNonUniqueTuples != 1) "s" else ""
+    status match {
+      case ConstraintSuccess => s"$columnNoun $columnsString $columnVerb a key."
+      case ConstraintFailure => s"$columnNoun $columnsString $columnVerb not a key ($numNonUniqueTuples non-unique tuple$pluralS)."
+    }
+  }
+
 }
