+++ src/test/scala/de/frosner/ddq/reporters/MarkdownReporterTest.scala
@@ -4,6 +4,7 @@ import java.io.{PrintStream, ByteArrayOutputStream}
 
 import de.frosner.ddq.constraints._
 import de.frosner.ddq.core._
+import de.frosner.ddq.testutils.{DummyConstraintResult, DummyConstraint}
 import org.apache.spark.sql.DataFrame
 import org.mockito.Mockito._
 import org.scalatest.mock.MockitoSugar
