+++ src/test/scala/de/frosner/ddq/testutils/DummyConstraint.scala
@@ -1,5 +1,6 @@
-package de.frosner.ddq.constraints
+package de.frosner.ddq.testutils
 
+import de.frosner.ddq.constraints.{Constraint, ConstraintStatus}
 import org.apache.spark.sql.DataFrame
 
 case class DummyConstraint(message: String, status: ConstraintStatus) extends Constraint {
