+++ src/test/scala/de/frosner/ddq/testutils/DummyConstraintResult.scala
@@ -1,4 +1,6 @@
-package de.frosner.ddq.constraints
+package de.frosner.ddq.testutils
+
+import de.frosner.ddq.constraints.{ConstraintResult, ConstraintStatus}
 
 case class DummyConstraintResult(constraint: DummyConstraint,
                                  message: String,
